feat/test: tag 클릭시 호출 기능 구현완료/ 테스트 완료

+++ src/main/java/com/book_everywhere/domain/book/BookRepository.java
@@ -2,5 +2,5 @@ package com.book_everywhere.domain.book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
 }
