+++ src/main/java/com/book_everywhere/domain/pin/PinRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface PinRepository extends JpaRepository<Pin, Integer> {
+public interface PinRepository extends JpaRepository<Pin, Long> {
     // 유저 아이디를 넣으면 어디에 방믄했는지 모두 가져옴 -> 나만의 지도 기능
     @Query("SELECT pin FROM Pin pin JOIN pin.visits visit WHERE visit.user.socialId = :socialId")
     List<Pin> mUserMap(@Param("socialId") Long socialId);
